+++ CHANGELOG.md
@@ -3,7 +3,10 @@
 **8.3.1**
 
 ```
-UPD: Remove github worflow
+UPD: Remove github workflow
+
+IMP: Longpolling print
+IMP: Terminal style
 
 FIX: Get initial information (username and version): Now work correctly with public users
 FIX: Command 'login' and 'logout': Reload page to get updated session information